% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{evaluate_results}
\alias{evaluate_results}
\title{Evaluate results: print, plot, and suggest best method}
\usage{
evaluate_results(res, metric = "RMSE", higher_better = FALSE)
}
\arguments{
\item{res}{The result object returned from evaluator()}

\item{metric}{Character, which metric to optimize (default = "RMSE")}

\item{higher_better}{Logical, whether higher values indicate better performance}
}
\value{
Prints a table, plots results, and returns the suggested method
}
\description{
Evaluate results: print, plot, and suggest best method
}
