% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{plot_density_all}
\alias{plot_density_all}
\title{Plot combined density comparisons for all columns and methods}
\usage{
plot_density_all(eval_list)

plot_density_all(eval_list)
}
\arguments{
\item{eval_list}{Output of get_eval_list(res)}
}
\value{
A faceted ggplot object showing all density comparisons
}
\description{
Plot combined density comparisons for all columns and methods

Plot consolidated density comparisons for all columns and methods
}
