% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{plot_density_per_column}
\alias{plot_density_per_column}
\title{Plot density comparisons of true vs imputed values for one column}
\usage{
plot_density_per_column(eval_list, col_name)
}
\arguments{
\item{eval_list}{Output of get_eval_list(res)}

\item{col_name}{Column to plot}
}
\description{
Plot density comparisons of true vs imputed values for one column
}
