% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{plot_metrics}
\alias{plot_metrics}
\title{Plot evaluation metrics}
\usage{
plot_metrics(x, metric = "RMSE")
}
\arguments{
\item{x}{A result list returned by evaluator()}

\item{metric}{Character, which metric to plot (e.g., "RMSE", "MAE", "R2", "Accuracy", "F1", "ALL")}
}
\value{
A ggplot2 object
}
\description{
Plot evaluation metrics (numeric + categorical)
}
