% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{print.evaluator}
\alias{print.evaluator}
\title{Print method for evaluator objects}
\usage{
\method{print}{evaluator}(x, ...)
}
\arguments{
\item{x}{An evaluator object}

\item{...}{further arguments}
}
\description{
Print method for evaluator objects
}
