% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{print.evaluator}
\alias{print.evaluator}
\title{print.evaluator}
\usage{
\method{print}{evaluator}(x, ...)
}
\arguments{
\item{x}{An evaluator object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for evaluator objects
}
\seealso{
vignette("imputetoolkit")
For a complete tutorial, see the package vignette:
\code{vignette("imputetoolkit")}
}
