% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{summary.evaluator}
\alias{summary.evaluator}
\title{summary.evaluator}
\usage{
\method{summary}{evaluator}(object, ...)
}
\arguments{
\item{object}{An evaluator object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for evaluator objects
}
\seealso{
\link[vignette]{imputetoolkit}
For a complete tutorial, see the package vignette:
\code{vignette("imputetoolkit")}
}
